nice: Refactor project file structure for better organization and maintainability
