+++ src/hooks/useOnline.js
@@ -1,8 +1,5 @@
 import { useEffect, useState } from "react";
 
-
-//removing event lissners is a good practice because it not stop after returning
-
 const useOnline=()=>{
 const [isOnline,setIsOnline]=useState(true);
 const onlineH=()=>{
